+++ slang/lib/builder/utils/regex_utils.dart
@@ -52,10 +52,10 @@ class RegexUtils {
   /// 1 - MyGeneric
   static RegExp genericRegex = RegExp(r'^List<((?:\w| |<|>)+)>$');
 
-  /// Matches a parameter hint in a key if it exists
-  /// greet(plural, param: gender)
-  /// 1 - plural, param: gender
-  static RegExp hintRegex = RegExp(r'^\w+\((.+)\)$');
+  /// Matches the modifier part in a key if it exists
+  /// greet(plural, param=gender)
+  /// 1 - plural, param=gender
+  static RegExp modifierRegex = RegExp(r'^\w+\((.+)\)$');
 
   static RegExp spaceRegex = RegExp(r'\s+');